LayoutTests/fast/forms/ValidityState-tooLong-textarea-expected.txt

Tests for tooLong flag with <textarea> elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textarea.validity.tooLong is false
PASS textarea.value.length is 5
PASS textarea.validity.tooLong is false
PASS textarea.value.length is 6
PASS textarea.validity.tooLong is false
PASS textarea.value.length is 4
PASS textarea.validity.tooLong is true
PASS textarea.validity.tooLong is false
PASS textarea.validity.tooLong is true
PASS successfullyParsed is true

TEST COMPLETE
